# Optimize PreComputeTupleReader