``` ini
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
AMD Ryzen 7 3700X, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=6.0.100-preview.3.21173.9
  [Host]     : .NET Core 5.0.2 (CoreCLR 5.0.220.61120, CoreFX 5.0.220.61120), X64 RyuJIT DEBUG
  DefaultJob : .NET Core 5.0.2 (CoreCLR 5.0.220.61120, CoreFX 5.0.220.61120), X64 RyuJIT
```
|   Method |         Mean |      Error |     StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |-------------:|-----------:|-----------:|-------:|------:|------:|----------:|
| New3Part |    78.389 ns |  3.2447 ns |  9.5672 ns | 0.0592 |     - |     - |     496 B |
| Old3Part | 1,998.145 ns | 16.6370 ns | 14.7483 ns | 0.1144 |     - |     - |     984 B |
| New1Part |     9.285 ns |  0.1842 ns |  0.1538 ns | 0.0076 |     - |     - |      64 B |
| Old1Part |   180.918 ns |  3.5238 ns |  3.2961 ns | 0.0076 |     - |     - |      64 B |

Tested against
```fsharp
let threePart = (("yup", 1s), 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, Some 12, 13, "nope",  17, 18, ValueSome 19)
let onePart = ("yup", 1s)
```

And I think that's the conclusion to optimizing the PreCompute family of methods 🥳 (as far as big gains like this go anyway).