@kerams 2021-04-06
Done. Yes, the implementation is even more complex than `compileTupleConstructor`, but I don't think it could be simplified much. As you say, there are more things the emitted code needs to do, and the fact that large tuples are recursive in nature does not help either.